+++ treecorr/ggcorrelation.py
@@ -411,7 +411,7 @@ class GGCorrelation(treecorr.BinnedCorr2):
         """
         self.logger.info('Reading GG correlations from %s',file_name)
 
-        data = treecorr.util.gen_read(file_name, file_type=file_type)
+        data, _ = treecorr.util.gen_read(file_name, file_type=file_type)
         self.rnom = data['R_nom']
         self.logr = numpy.log(data['R_nom'])
         self.meanr = data['meanR']