+++ treecorr/kgcorrelation.py
@@ -355,7 +355,7 @@ class KGCorrelation(treecorr.BinnedCorr2):
         """
         self.logger.info('Reading KG correlations from %s',file_name)
 
-        data = treecorr.util.gen_read(file_name, file_type=file_type)
+        data, _ = treecorr.util.gen_read(file_name, file_type=file_type)
         self.rnom = data['R_nom']
         self.logr = numpy.log(self.rnom)
         self.meanr = data['meanR']