+++ treecorr/ngcorrelation.py
@@ -382,7 +382,7 @@ class NGCorrelation(treecorr.BinnedCorr2):
         """
         self.logger.info('Reading NG correlations from %s',file_name)
 
-        data = treecorr.util.gen_read(file_name, file_type=file_type)
+        data, _ = treecorr.util.gen_read(file_name, file_type=file_type)
         self.rnom = data['R_nom']
         self.logr = numpy.log(self.rnom)
         self.meanr = data['meanR']